refactor: Move duplicated code to reusable hooks
Refs PR #174
