+++ web/src/components/Market/Outcomes.tsx
@@ -7,6 +7,8 @@ import { useModal } from "@/hooks/useModal";
 import { useSearchParams } from "@/hooks/useSearchParams";
 import { useTokenBalances } from "@/hooks/useTokenBalance";
 import { useTokensInfo } from "@/hooks/useTokenInfo";
+import { useWinningOutcomes } from "@/hooks/useWinningOutcomes";
+import { useSortedOutcomes } from "@/hooks/useSortedOutcomes";
 import { SUPPORTED_CHAINS, SupportedChain } from "@/lib/chains";
 import { COLLATERAL_TOKENS, SWAPR_CONFIG, getFarmingUrl, getLiquidityUrl } from "@/lib/config";
 import { CheckCircleIcon, EtherscanIcon, QuestionIcon, RightArrow } from "@/lib/icons";
@@ -17,7 +19,7 @@ import { displayBalance, isUndefined, toSnakeCase } from "@/lib/utils";
 import { config } from "@/wagmi";
 import { getConnectorClient } from "@wagmi/core";
 import clsx from "clsx";
-import { useEffect, useMemo } from "react";
+import { useEffect } from "react";
 import { RpcError, zeroAddress } from "viem";
 import { watchAsset } from "viem/actions";
 import { useAccount } from "wagmi";
@@ -25,23 +27,11 @@ import { Alert } from "../Alert";
 import Button from "../Form/Button";
 import { Spinner } from "../Spinner";
 import { OutcomeImage } from "./OutcomeImage";
-import { RouterAbi } from "@/abi/RouterAbi";
-import { useQuery } from "@tanstack/react-query";
-import { readContract } from "@wagmi/core";
 import { Address } from "viem";
-import { getRouterAddress } from "@/lib/config";
-import { MarketStatus, useMarketStatus } from "@/hooks/useMarketStatus";
-import { INVALID_RESULT_OUTCOME_TEXT } from "@/lib/utils";
-
+import { useMarketStatus } from "@/hooks/useMarketStatus";
 interface PositionsProps {
   market: Market;
   images?: string[];
-}
-
-type OutcomeWithOdds = {
-  odd: number;
-  i: number;
-  isWinning?: boolean;
 }
 
 function poolRewardsInfo(poolIncentive: PoolIncentive) {
@@ -214,58 +204,10 @@ export function Outcomes({ market, images }: PositionsProps) {
   const { data: pools = [] } = useMarketPools(market);
   const { Modal, openModal, closeModal } = useModal("liquidity-modal");
   const blockExplorerUrl = SUPPORTED_CHAINS[market.chainId].blockExplorers?.default?.url;
-
-  const routerAddress = getRouterAddress(market.chainId);
   const { data: marketStatus } = useMarketStatus(market);
-  const winningOutcomes = useQuery({
-    queryKey: ["winningOutcomes", market.conditionId, routerAddress],
-    enabled: marketStatus === MarketStatus.CLOSED && !!routerAddress,
-    queryFn: async () => {
-      return await readContract(config, {
-        abi: RouterAbi,
-        address: routerAddress as Address,
-        functionName: "getWinningOutcomes",
-        args: [market.conditionId],
-        chainId: market.chainId
-      });
-    }
-  });
+  const { data: winningOutcomes } = useWinningOutcomes(market.conditionId as Address, market.chainId, marketStatus);
+  const { data: indexesOrderedByOdds } = useSortedOutcomes(market, marketStatus);
 
-  const indexesOrderedByOdds = useMemo(() => {
-    if (oddsPending || odds.length === 0) {
-      return null;
-    }
-    const invalidIndex = market.outcomes.findIndex(outcome => outcome === INVALID_RESULT_OUTCOME_TEXT);
-
-    if (!winningOutcomes.data && marketStatus === MarketStatus.CLOSED) {
-      const otherIndexes = odds
-        .map((odd, i) => ({ odd, i }))
-        .filter(({ i }) => i !== invalidIndex)
-        .sort((a, b) => b.odd - a.odd)
-        .map(obj => obj.i);
-
-      return [invalidIndex, ...otherIndexes];
-    }
-
-
-    const winningIndexes: OutcomeWithOdds[] = [];
-    const nonWinningIndexes: OutcomeWithOdds[] = [];
-
-    odds.forEach((odd, i) => {
-      if (winningOutcomes.data?.[i] === true) {
-        winningIndexes.push({ odd, i });
-      } else {
-        nonWinningIndexes.push({ odd, i });
-      }
-    });
-
-    // Sort each group by odds
-    const sortedWinning = winningIndexes.sort((a, b) => b.odd - a.odd).map(obj => obj.i);
-    const sortedNonWinning = nonWinningIndexes.sort((a, b) => b.odd - a.odd).map(obj => obj.i);
-
-    return [...sortedWinning, ...sortedNonWinning];
-  }, [odds, winningOutcomes.data, market.outcomes, marketStatus]);
-   
 
   useEffect(() => {
     if (!searchParams.get("outcome") && indexesOrderedByOdds) {
@@ -360,7 +302,7 @@ export function Outcomes({ market, images }: PositionsProps) {
                         <QuestionIcon fill="#9747FF" />
                       </span>
                     )}
-                      {winningOutcomes.data?.[i] === true &&  <CheckCircleIcon className="text-success-primary" />}
+                      {winningOutcomes?.[i] === true &&  <CheckCircleIcon className="text-success-primary" />}
                   </div>
                   <div className="text-[12px] text-black-secondary">
                     {balances && balances[i] > 0n && (
